requirements.txt
streamlit
torch
torchvision
fastai
Pillow
scikit-image
numpy